requirements.txt
fastapi
uvicorn
psycopg2-binary
sentence-transformers
faiss-cpu
streamlit
requests
sqlparse